+++ src/components/Searchbar.js
@@ -1,13 +1,10 @@
 import React from "react";
-// import { Button, ButtonGroup } from "react-bootstrap";
 import "../index.css";
 
 function Searchbar(props) {
-    
+
     const handleChange = props.handleChange
     
-
-
 return (
     <div className="Searchbar">
         <form className="form-searchbar">
