+++ src/components/useFetch.js
@@ -1,40 +1,36 @@
 import { useState, useEffect } from "react";
 
-const useFetch = (url) => {
+const useFetch = (url, numItems) => {
   const [data, setData] = useState(null)
-    const [loading, setLoading] = useState(true)
+  const [loading, setLoading] = useState(true)
   const [error, setError] = useState(null)
+  var fetchUrl = url+'?limit='+numItems
+  console.log(fetchUrl)
   
 
-    const getData = async (url) => {
-         console.log("getData")
-try {
-const response = await fetch(url);
-if (!response.ok) {
-    throw new Error(
-    `This is an HTTP error: The status is ${response.status}`
-    );
-}
-let actualData = await response.json();
-const myData = actualData
-setData(actualData)
-
-setError(null);
-} catch (err) {
-    setError(err.message);
-setData(null);
-} finally {
-    setLoading(false);
-}
+  const getData = async (fetchUrl) => {
+    try {
+    const response = await fetch(fetchUrl);
+    if (!response.ok) {
+        throw new Error(
+        `This is an HTTP error: The status is ${response.status}`);
     }
+    let actualData = await response.json();
+    setData(actualData)
+    setError(null);
+    } catch (err) {
+        setError(err.message);
+    setData(null);
+    } finally {
+        setLoading(false);
+    }
+  }
 
-
-console.log("component")
   useEffect(() => {
-getData(url)
+    getData(fetchUrl)
   }, []);
 
-  return { data, loading, error }
+  return { data, loading, error, setData }
 };
 
 export default useFetch;
