+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
 import App from './App';
-import ViewProducts from './views/ViewProducts'
+import ViewShop from './views/ViewShop'
 import ViewChat from './views/ViewChat'
 import ViewContact from './views/ViewContact'
 import ViewLogin from './views/ViewLogin'
@@ -16,7 +16,7 @@ root.render(
   <BrowserRouter>
     <Routes>
       <Route exact path="/" element={<App />}/>
-      <Route path="/:id" element={<ViewProducts />}/>
+      <Route path="/:id" element={<ViewShop />}/>
       <Route path="/chat" element={<ViewChat />} />
       <Route path="/contact" element={<ViewContact />} />
       <Route path="/login" element={<ViewLogin />} />
